Refactor ChatPanel lifecycle tests
